refactor: leave the moved functions where they were
recreateUserHeaders was moved to utils.ts, and processRawSteps and
deduplicateSteps to a new raw-steps.ts, to break import cycles that the
first version of this feature created: workflow-requests.ts reached into
the discovery replay, and needed the raw step parsing for the call result
discovery. Neither exists any more, so neither move is needed, and the
diff was carrying three relocated functions for nothing.

Put them back. utils.ts is identical to main again, raw-steps.ts is gone,
and nothing in this PR relocates a function. The functions themselves are
unchanged, apart from the comment in processRawSteps which now names the
stepConfig call type its filter also skips.

Fewer import cycles than before the move, not more: madge counts 33 where
the moved layout had 34.
