+++ src/utils.ts
@@ -1,7 +1,6 @@
 import { WorkflowError, WorkflowNonRetryableError } from "./error";
 import type { FlowControl } from "@upstash/qstash";
 import { WorkflowClient } from "./types";
-import { WORKFLOW_LABEL_HEADER } from "./constants";
 
 const NANOID_CHARS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_";
 const NANOID_LENGTH = 21;
@@ -140,39 +139,3 @@ export function stringifyBody(body: unknown): string | undefined {
   }
   return JSON.stringify(body);
 }
-
-/**
- * Removes headers starting with `Upstash-Workflow-` from the headers
- *
- * @param headers incoming headers
- * @returns headers with `Upstash-Workflow-` headers removed
- */
-export const recreateUserHeaders = (headers: Headers): Headers => {
-  const filteredHeaders = new Headers();
-
-  const pairs = headers.entries() as unknown as [string, string][];
-  for (const [header, value] of pairs) {
-    const headerLowerCase = header.toLowerCase();
-
-    const isUserHeader =
-      (headerLowerCase !== "upstash-region" &&
-        !headerLowerCase.startsWith("upstash-workflow-") &&
-        // https://vercel.com/docs/edge-network/headers/request-headers#x-vercel-id
-        !headerLowerCase.startsWith("x-vercel-") &&
-        !headerLowerCase.startsWith("x-forwarded-") &&
-        // https://blog.cloudflare.com/preventing-request-loops-using-cdn-loop/
-        headerLowerCase !== "cf-connecting-ip" &&
-        headerLowerCase !== "cdn-loop" &&
-        headerLowerCase !== "cf-ew-via" &&
-        headerLowerCase !== "cf-ray" &&
-        // For Render https://render.com
-        headerLowerCase !== "render-proxy-ttl") ||
-      headerLowerCase === WORKFLOW_LABEL_HEADER.toLocaleLowerCase();
-
-    if (isUserHeader) {
-      filteredHeaders.append(header, value);
-    }
-  }
-
-  return filteredHeaders as Headers;
-};
